refactor docker package
